requirements.txt
asgiref==3.3.1
gunicorn==20.04
Django==3.1.4
django-crispy-forms==1.10.0
joblib==1.0.0
numpy==1.19.5
pytz==2020.5
scikit-learn==0.24.0
scipy==1.6.0
sqlparse==0.4.1
threadpoolctl==2.1.0
dj-database-url-0.5.0 
django-heroku-0.3.1 
psycopg2-2.8.6
whitenoise-5.2.0